E_learning_platform/requirements.txt
aioredis==1.3.1
asgiref==3.6.0
astroid==2.15.4
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==40.0.2
daphne==3.0.2
Deprecated==1.2.13
dill==0.3.6
Django==4.1
django-braces==1.15.0
django-debug-toolbar==3.6.0
django-embed-video==1.4.4
django-redisboard==8.3.0
djangorestframework==3.13.1
hiredis==2.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
msgpack==1.0.5
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pylint==2.17.3
pymemcache==3.5.2
pyOpenSSL==23.1.1
pytz==2023.3
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
zope.interface==6.0
